+++ src/trees/MWTree.cpp
@@ -196,7 +196,7 @@ template <int D, typename T> void MWTree<D, T>::mwTransformDown(bool overwrite) 
     tree_utils::make_node_table(*this, nodeTable);
 #pragma omp parallel shared(nodeTable) num_threads(mrcpp_get_num_threads())
     {
-        for (int n = 0; n < nodeTable.size(); n++) {
+        for (size_t n = 0; n < nodeTable.size(); n++) {
             int n_nodes = nodeTable[n].size();
 #pragma omp for schedule(guided)
             for (int i = 0; i < n_nodes; i++) {
@@ -284,9 +284,9 @@ template <int D, typename T> void MWTree<D, T>::decrementNodeCount(int scale) {
 template <int D, typename T> int MWTree<D, T>::getNNodesAtDepth(int depth) const {
     int N = 0;
     if (depth < 0) {
-        if (this->nodesAtNegativeDepth.size() >= -depth) N = this->nodesAtNegativeDepth[-depth];
+        if (this->nodesAtNegativeDepth.size() >= static_cast<size_t>(-depth)) N = this->nodesAtNegativeDepth[-depth];
     } else {
-        if (this->nodesAtDepth.size() > depth) N = this->nodesAtDepth[depth];
+        if (this->nodesAtDepth.size() > static_cast<size_t>(depth)) N = this->nodesAtDepth[depth];
     }
     return N;
 }
@@ -523,8 +523,8 @@ template <int D, typename T> std::ostream &MWTree<D, T>::print(std::ostream &o) 
     o << "  nodes: " << this->getNNodes() << std::endl;
     o << "  endNodes: " << this->endNodeTable.size() << std::endl;
     o << "  nodes per scale: " << std::endl;
-    for (int i = this->nodesAtNegativeDepth.size() - 1; i >= 0; i--) { o << "    scale=" << -(i + this->getRootScale() + 1) << "  nodes=" << this->nodesAtNegativeDepth[i] << std::endl; }
-    for (int i = 0; i < this->nodesAtDepth.size(); i++) { o << "    scale=" << i + this->getRootScale() << "  nodes=" << this->nodesAtDepth[i] << std::endl; }
+    for (size_t i = this->nodesAtNegativeDepth.size() - 1; i >= 0; i--) { o << "    scale=" << -(i + this->getRootScale() + 1) << "  nodes=" << this->nodesAtNegativeDepth[i] << std::endl; }
+    for (size_t i = 0; i < this->nodesAtDepth.size(); i++) { o << "    scale=" << i + this->getRootScale() << "  nodes=" << this->nodesAtDepth[i] << std::endl; }
     return o;
 }
